+++ packages/blocks/src/store/test/private-selectors.js
@@ -63,10 +63,18 @@ describe( 'private selectors', () => {
 				'padding',
 				'contentSize',
 				'wideSize',
-				'blockGap',
-				'textAlign',
 				'textDecoration',
 				'textShadow',
+				'backgroundGradient',
+				'backgroundImage',
+				'backgroundSize',
+				'borderColor',
+				'borderRadius',
+				'borderStyle',
+				'borderWidth',
+				'margin',
+				'shadow',
+				'writingMode',
 			] );
 		} );
 
@@ -92,11 +100,19 @@ describe( 'private selectors', () => {
 				'padding',
 				'contentSize',
 				'wideSize',
-				'blockGap',
-				'textAlign',
 				'textTransform',
 				'textShadow',
 				'letterSpacing',
+				'backgroundGradient',
+				'backgroundImage',
+				'backgroundSize',
+				'borderColor',
+				'borderRadius',
+				'borderStyle',
+				'borderWidth',
+				'margin',
+				'shadow',
+				'writingMode',
 			] );
 		} );